feat(timeline): mount the rotation view behind ?newTimeline
Both views are built and one is shown, so the default path is unchanged
and the new one can be driven in a browser. Wiring it is what found the
defects below; none of them were visible to the type checker.

Three came out of a constructor that had never executed.

The window mounted zero rows. The scroller is content-sized, so it had no
height until the spacers were written, but the window pass bails on a zero
clientHeight and so never wrote them. Measured clientHeight 0, scrollHeight
0, no rows. invalidate() now sizes the bottom spacer to the full content
height before the first pass.

Registering the new stylesheet broke the old view. Both put .rotation-row
on their rows, so the later import won and legacy rows picked up flex, a
1px height and layout containment. The two colliding selectors are scoped
to .rotation-pane, which makes the alphabetical import order load-bearing:
timeline_rotation must stay after timeline.

The stuck observer used the APL bar's root: parent, which makes the root
rect degenerate so the ratio is always below one. It now observes against
the viewport with a bottom margin, like sticky_toolbar. Worth recording
while here: the APL bar itself is permanently stuck, because its
classList.toggle('stuck') has no second argument and so flips rather than
tracking the observation.

The row visibility UI becomes RotationFloatingActionBar, matching the APL
bar in shape and in chrome: transparent and borderless in flow, growing
padding, border and an opaque background only once pinned. It cannot use
that partial's --apl-spacer, which is scoped to the APL picker subtree and
would resolve to nothing here, so the padding is var(--spacer-2). Chips
reuse saved-data-set-chip badge rounded-pill with .active for shown rather
than a private pill style.

The bar overlays rather than resizing the timeline. The scroller is
flex: 0 0 auto, so an in-flow bar cannot shrink it -- clientHeight measures
700 with the bar absent, present, collapsed and expanded -- and the panel
grows upward from the bar's top edge. The scroller's bottom padding covers
the bar's stuck height, not its resting height, so the last row still
clears it once the chrome appears.
