+++ schemas/translation.schema.json
@@ -7056,10 +7056,29 @@
 									},
 									"additionalProperties": false,
 									"required": ["dps", "dtps", "amount", "player_damage", "player_damage_taken", "before", "after", "threat", "active_auras"]
+								},
+								"floatingActionBar": {
+									"type": "object",
+									"properties": {
+										"allShown": {
+											"type": "string"
+										},
+										"hidden": {
+											"type": "string"
+										},
+										"showAll": {
+											"type": "string"
+										},
+										"toggle": {
+											"type": "string"
+										}
+									},
+									"additionalProperties": false,
+									"required": ["allShown", "hidden", "showAll", "toggle"]
 								}
 							},
 							"additionalProperties": false,
-							"required": ["disclaimer", "note", "chart_types", "chart_options", "tooltips"]
+							"required": ["disclaimer", "note", "chart_types", "chart_options", "tooltips", "floatingActionBar"]
 						},
 						"logs": {
 							"type": "object",
